+++ packages/integration-test/importNamed.test.mjs
@@ -1,6 +1,6 @@
 /** @jest-environment jsdom */
 
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@testing-library/react';
 import useRefFrom from 'use-ref-from/useRefFrom';
 
 test('simple scenario', () => {