+++ packages/use-ref-from/src/useRefFrom.spec.ts
@@ -1,4 +1,6 @@
-import { renderHook } from '@testing-library/react-hooks';
+/** @jest-environment jsdom */
+
+import { renderHook } from '@testing-library/react';
 import { useCallback } from 'react';
 import useRefFrom from './useRefFrom';